A .NET Core-compatible (netstandard1.3 + .NET 4) spike.
