+++ .gitattributes
@@ -1,0 +1,3 @@
+# Auto detect text files and perform LF normalization
+
+* text=auto
