+++ .gitignore
@@ -22,3 +22,4 @@ _ReSharper*/
 .vs/
 Blackstar.1.0.0.0.nupkg
 nuget.exe
+artifacts/
