--- Blackstar/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net40" />
-  <package id="Microsoft.Bcl.Async" version="1.0.168" targetFramework="net40" />
-  <package id="Microsoft.Bcl.Build" version="1.0.14" targetFramework="net40" />
-  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="net40" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net40" />
-</packages>
