--- Tests/AssertionExtensions.cs
@@ -1,16 +1,0 @@
-﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Tests
-{
-    public static class AssertionExtensions
-    {
-        public static void UriEqual(string expected, Uri actual)
-        {
-            Assert.AreEqual(expected, actual.ToString());
-        }
-    }
-}
