+++ src/Blackstar/BlackstarClient.cs
