+++ src/Blackstar/ContentChunk.cs
