+++ src/Blackstar/project.json
@@ -1,0 +1,41 @@
+﻿{
+  "version": "1.2.0-*",
+
+  "description": "This project is the .NET client for Blackstar CMS",
+  "authors": [ "Eclipse Web Solutions Pty Ltd" ],
+
+  "packOptions": {
+    "tags": [ "cms" ],
+    "projectUrl": "https://github.com/Blackstar-CMS/dotnet-client",
+    "licenseUrl": "http://choosealicense.com/licenses/mit/",
+    "iconUrl": "http://blackstarcms.net/images/mac2.png"
+  },
+
+  "buildOptions": {
+    "warningsAsErrors": true
+  },
+
+  "dependencies": {
+    "Newtonsoft.Json": "9.0.1"
+  },
+
+  "frameworks": {
+    "net4.0": {
+      "dependencies": {
+        "Microsoft.Net.Http": "2.2.29",
+        "Microsoft.Bcl.Async": "1.0.168"
+      }
+    },
+    "netstandard1.3": {
+      "dependencies": {
+        "Microsoft.CSharp": "4.0.1",
+        "System.Collections": "4.0.11",
+        "System.Dynamic.Runtime": "4.0.11",
+        "System.Globalization": "4.0.11",
+        "System.Linq": "4.1.0",
+        "System.Net.Http":  "4.1.0",
+        "System.Runtime": "4.1.0"
+      }
+    }
+  }
+}
