+++ test/Blackstar.Tests/AssertionExtensions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Xunit;
+
+namespace Tests
+{
+    public static class AssertionExtensions
+    {
+        public static void UriEqual(string expected, Uri actual)
+        {
+            Assert.Equal(expected, actual.ToString());
+        }
+    }
+}
