+++ test/Blackstar.Tests/project.json
@@ -1,0 +1,25 @@
+﻿{
+  "testRunner": "xunit",
+
+  "dependencies": {
+    "Blackstar": { "target": "project" },
+    "xunit": "2.1.0",
+    "dotnet-test-xunit": "1.0.0-rc2-build10025"
+  },
+
+  "frameworks": {
+    "net4.5.2": { },
+    "netcoreapp1.0": {
+      "dependencies": {
+        "Microsoft.NETCore.App": {
+          "type": "platform",
+          "version": "1.0.0"
+        }
+      },
+      "imports": [
+        "dnxcore50",
+        "portable-net45+win8"
+      ]
+    }
+  }
+}
